Relocate default impl off of API class
